Fix windows build.

+++ src/eepp/ui/doc/textdocument.cpp
@@ -3024,7 +3024,7 @@ void TextDocument::setDirtyUntilSave() {
 }
 
 static size_t guessFileSize( const TextDocument* doc ) {
-	Int64 maxLines = std::min( 101ul, doc->linesCount() );
+	Int64 maxLines = std::min( (Int64)101, (Int64)doc->linesCount() );
 	Int64 totalSize = 0;
 	for ( Int64 i = 0; i < maxLines; i++ )
 		totalSize += doc->line( i ).size();
